requirements-cpu.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aenum==3.1.15
    # via lightly
aiohttp==3.9.3
    # via
    #   aiohttp-retry
    #   fsspec
aiohttp-retry==2.8.3
    # via
    #   dvc-http
    #   scmrepo
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
appdirs==1.4.4
    # via
    #   iterative-telemetry
    #   wandb
async-timeout==4.0.3
    # via aiohttp
asyncssh==2.14.2
    # via scmrepo
atpublic==4.0
    # via flufl-lock
attrs==23.2.0
    # via
    #   aiohttp
    #   dvc-data
    #   sqltrie
billiard==4.2.0
    # via celery
celery==5.3.6
    # via dvc-task
certifi==2024.2.2
    # via
    #   lightly
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pygit2
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   typer
    #   wandb
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via dvc
configobj==5.0.8
    # via dvc
cryptography==42.0.5
    # via asyncssh
cython==3.0.8
    # via -r requirements.in
dictdiffer==0.9.0
    # via dvc-data
diskcache==5.6.3
    # via dvc-data
distro==1.9.0
    # via
    #   dvc
    #   iterative-telemetry
docker-pycreds==0.4.0
    # via wandb
dpath==2.1.6
    # via dvc
dulwich==0.21.7
    # via
    #   dvc-studio-client
    #   scmrepo
dvc==3.48.1
    # via -r requirements.in
dvc-data==3.14.0
    # via dvc
dvc-http==2.32.0
    # via dvc
dvc-objects==5.1.0
    # via dvc-data
dvc-render==1.0.1
    # via dvc
dvc-studio-client==0.20.0
    # via dvc
dvc-task==0.3.0
    # via dvc
entrypoints==0.4
    # via gto
filelock==3.13.1
    # via
    #   iterative-telemetry
    #   torch
flatten-dict==0.4.2
    # via dvc
flufl-lock==7.1.1
    # via dvc
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http,tqdm]==2024.2.0
    # via
    #   dvc-data
    #   dvc-http
    #   dvc-objects
    #   lightning
    #   lightning-lite
    #   pytorch-lightning
    #   scmrepo
    #   torch
funcy==2.0
    # via
    #   dvc
    #   dvc-data
    #   dvc-objects
    #   dvc-task
    #   gto
    #   scmrepo
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via
    #   scmrepo
    #   wandb
grandalf==0.8
    # via dvc
gto==1.7.0
    # via dvc
hydra-core==1.3.2
    # via
    #   -r requirements.in
    #   dvc
    #   lightly
idna==3.6
    # via
    #   requests
    #   yarl
iterative-telemetry==0.0.8
    # via dvc
jinja2==3.1.3
    # via torch
joblib==1.3.2
    # via scikit-learn
kombu==5.3.5
    # via
    #   celery
    #   dvc-task
lightly==1.5.0
    # via -r requirements.in
lightly-utils==0.0.2
    # via lightly
lightning==2.2.1
    # via -r requirements.in
lightning-lite==1.8.6
    # via -r requirements.in
lightning-utilities==0.10.1
    # via
    #   lightning
    #   lightning-lite
    #   pytorch-lightning
    #   torchmetrics
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via
    #   dvc
    #   torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   lightly
    #   lightly-utils
    #   lightning
    #   lightning-lite
    #   pandas
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   torchmetrics
    #   torchvision
omegaconf==2.3.0
    # via
    #   -r requirements.in
    #   hydra-core
openood @ git+https://github.com/djaniak/OpenOOD.git
    # via
    #   manual entry
orjson==3.9.15
    # via sqltrie
packaging==23.2
    # via
    #   dvc
    #   hydra-core
    #   lightning
    #   lightning-lite
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
pandas==2.2.1
    # via -r requirements.in
pathspec==0.12.1
    # via
    #   dvc
    #   scmrepo
pillow==10.2.0
    # via
    #   -r requirements.in
    #   lightly-utils
    #   torchvision
platformdirs==3.11.0
    # via dvc
prompt-toolkit==3.0.43
    # via click-repl
protobuf==4.25.3
    # via wandb
psutil==5.9.8
    # via
    #   dvc
    #   flufl-lock
    #   wandb
pycparser==2.21
    # via cffi
pydantic==1.10.14
    # via
    #   gto
    #   lightly
pydot==2.0.0
    # via dvc
pygit2==1.14.1
    # via scmrepo
pygments==2.17.2
    # via rich
pygtrie==2.5.0
    # via
    #   dvc
    #   dvc-data
    #   scmrepo
    #   sqltrie
pyparsing==3.1.2
    # via
    #   dvc
    #   grandalf
    #   pydot
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   lightly
    #   pandas
pytorch-lightning==2.2.1
    # via
    #   -r requirements.in
    #   lightly
    #   lightning
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   lightning
    #   omegaconf
    #   pytorch-lightning
    #   wandb
requests==2.31.0
    # via
    #   dvc
    #   dvc-studio-client
    #   iterative-telemetry
    #   lightly
    #   wandb
rich==13.7.1
    # via
    #   dvc
    #   gto
ruamel-yaml==0.18.6
    # via
    #   dvc
    #   gto
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
scikit-learn==1.4.1.post1
    # via -r requirements.in
scipy==1.12.0
    # via scikit-learn
scmrepo==3.2.0
    # via
    #   dvc
    #   gto
semver==3.0.2
    # via gto
sentry-sdk==1.40.6
    # via wandb
setproctitle==1.3.3
    # via wandb
shortuuid==1.0.12
    # via
    #   dvc
    #   dvc-task
shtab==1.7.0
    # via dvc
six==1.16.0
    # via
    #   configobj
    #   docker-pycreds
    #   flatten-dict
    #   lightly
    #   python-dateutil
smmap==5.0.1
    # via gitdb
sqltrie==0.11.0
    # via dvc-data
sympy==1.12
    # via torch
tabulate==0.9.0
    # via
    #   dvc
    #   gto
threadpoolctl==3.3.0
    # via scikit-learn
tomlkit==0.12.4
    # via dvc
torch==2.2.1
    # via
    #   -r requirements.in
    #   lightly
    #   lightning
    #   lightning-lite
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==1.3.1
    # via
    #   -r requirements.in
    #   lightning
    #   pytorch-lightning
torchvision==0.17.1
    # via
    #   -r requirements.in
    #   lightly
tqdm==4.66.2
    # via
    #   -r requirements.in
    #   dvc
    #   dvc-data
    #   fsspec
    #   lightly
    #   lightning
    #   pytorch-lightning
    #   scmrepo
typer==0.9.0
    # via
    #   -r requirements.in
    #   gto
typing-extensions==4.10.0
    # via
    #   -r requirements.in
    #   asyncssh
    #   lightning
    #   lightning-lite
    #   lightning-utilities
    #   pydantic
    #   pytorch-lightning
    #   torch
    #   typer
tzdata==2024.1
    # via
    #   celery
    #   pandas
urllib3==2.2.1
    # via
    #   dulwich
    #   lightly
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
voluptuous==0.14.2
    # via
    #   dvc
    #   dvc-studio-client
wandb==0.16.4
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.9.4
    # via aiohttp
zc-lockfile==3.0.post1
    # via dvc

# The following packages are considered to be unsafe in a requirements file:
# setuptools
